Removed new iOS 7 text sizing and just set some default sizes instead, modified drawing routine slightly

+++ JA Battery/JABBatteryView.h
@@ -11,6 +11,6 @@
 
 @interface JABBatteryView : UIView
 
-@property (nonatomic) NSInteger batteryPercentage;
+@property (nonatomic) float batteryPercentage;
 
 @end
